backend/python-ml-service/app/services/avatar_service.py
"""
Avatar Service - Basic Implementation
Handles avatar generation and processing without heavy ML dependencies
"""

import uuid
import asyncio
import logging
from typing import Dict, Any, Optional
from fastapi import UploadFile
import cv2
import numpy as np
from PIL import Image
import io

from app.core.config import settings
from app.core.exceptions import AvatarProcessingException, InvalidImageException

logger = logging.getLogger(__name__)


class AvatarService:
    """
    Avatar processing service
    Basic implementation without TensorFlow - focuses on image processing
    """
    
    def __init__(self):
        self.processing_status = {}  # In-memory storage for demo
    
    async def start_avatar_generation(
        self,
        user_id: int,
        avatar_name: str,
        generation_method: str,
        image_file: Optional[UploadFile] = None,
        measurements: Optional[Dict[str, float]] = None,
        background_tasks = None
    ) -> Dict[str, Any]:
        """
        Start avatar generation process
        """
        try:
            processing_id = str(uuid.uuid4())
            
            # Initialize processing status
            self.processing_status[processing_id] = {
                "status": "pending",
                "progress": 0,
                "message": "Starting avatar generation...",
                "user_id": user_id,
                "avatar_name": avatar_name,
                "generation_method": generation_method
            }
            
            if generation_method == "manual_entry":
                # Handle manual measurements
                result = await self._process_manual_measurements(
                    processing_id, measurements
                )
            elif image_file:
                # Handle image-based generation
                result = await self._process_image_avatar(
                    processing_id, image_file
                )
            else:
                raise AvatarProcessingException("Invalid generation method or missing data")
            
            return {
                "processing_id": processing_id,
                "estimated_time": 30  # seconds
            }
            
        except Exception as e:
            logger.error(f"Error starting avatar generation: {str(e)}")
            raise AvatarProcessingException(f"Failed to start avatar generation: {str(e)}")
    
    async def _process_manual_measurements(
        self, 
        processing_id: str, 
        measurements: Dict[str, float]
    ) -> Dict[str, Any]:
        """Process manually entered measurements"""
        
        # Update status
        self.processing_status[processing_id].update({
            "status": "processing",
            "progress": 50,
            "message": "Processing manual measurements..."
        })
        
        # Simulate processing time
        await asyncio.sleep(2)
        
        # Basic validation and avatar creation
        avatar_data = {
            "mesh_url": f"/avatars/{processing_id}/mesh.obj",
            "texture_url": f"/avatars/{processing_id}/texture.png",
            "measurements": measurements,
            "generation_method": "manual_entry"
        }
        
        # Complete processing
        self.processing_status[processing_id].update({
            "status": "completed",
            "progress": 100,
            "message": "Avatar generated successfully",
            "result": avatar_data
        })
        
        return avatar_data
    
    async def _process_image_avatar(
        self, 
        processing_id: str, 
        image_file: UploadFile
    ) -> Dict[str, Any]:
        """Process image-based avatar generation"""
        
        try:
            # Update status
            self.processing_status[processing_id].update({
                "status": "processing",
                "progress": 10,
                "message": "Processing uploaded image..."
            })
            
            # Read and validate image
            image_data = await self._read_and_validate_image(image_file)
            
            # Update progress
            self.processing_status[processing_id].update({
                "progress": 30,
                "message": "Extracting features from image..."
            })
            
            # Basic image processing (without ML for now)
            measurements = await self._extract_basic_measurements(image_data)
            
            # Update progress
            self.processing_status[processing_id].update({
                "progress": 70,
                "message": "Generating 3D avatar..."
            })
            
            # Simulate avatar generation
            await asyncio.sleep(3)
            
            avatar_data = {
                "mesh_url": f"/avatars/{processing_id}/mesh.obj",
                "texture_url": f"/avatars/{processing_id}/texture.png",
                "measurements": measurements,
                "generation_method": "selfie",
                "confidence_score": 0.85
            }
            
            # Complete processing
            self.processing_status[processing_id].update({
                "status": "completed",
                "progress": 100,
                "message": "Avatar generated successfully from image",
                "result": avatar_data
            })
            
            return avatar_data
            
        except Exception as e:
            # Mark as failed
            self.processing_status[processing_id].update({
                "status": "failed",
                "progress": 0,
                "message": f"Avatar generation failed: {str(e)}",
                "error_details": str(e)
            })
            raise
    
    async def _read_and_validate_image(self, image_file: UploadFile) -> np.ndarray:
        """Read and validate uploaded image"""
        
        try:
            # Read image data
            image_bytes = await image_file.read()
            
            # Validate file size
            if len(image_bytes) > settings.MAX_FILE_SIZE:
                raise InvalidImageException("Image file too large")
            
            # Convert to PIL Image
            pil_image = Image.open(io.BytesIO(image_bytes))
            
            # Convert to OpenCV format
            image_array = np.array(pil_image)
            
            # Convert RGB to BGR (OpenCV format)
            if len(image_array.shape) == 3:
                image_array = cv2.cvtColor(image_array, cv2.COLOR_RGB2BGR)
            
            return image_array
            
        except Exception as e:
            raise InvalidImageException(f"Invalid image format: {str(e)}")
    
    async def _extract_basic_measurements(self, image: np.ndarray) -> Dict[str, float]:
        """
        Extract basic measurements from image
        This is a simplified version without ML models
        """
        
        # Get image dimensions for basic calculations
        height, width = image.shape[:2]
        
        # Simulate basic measurements (in a real implementation, this would use ML models)
        measurements = {
            "height_cm": 170.0,  # Default height
            "chest_cm": 90.0,
            "waist_cm": 75.0,
            "hip_cm": 95.0,
            "shoulder_width_cm": 45.0,
            "arm_length_cm": 60.0,
            "leg_length_cm": 80.0,
            "neck_cm": 35.0,
            "confidence": {
                "height": 0.7,
                "chest": 0.6,
                "waist": 0.6,
                "hip": 0.6,
                "overall": 0.65
            }
        }
        
        logger.info(f"Extracted basic measurements from image {width}x{height}")
        
        return measurements
    
    async def get_processing_status(self, processing_id: str) -> Dict[str, Any]:
        """Get the current status of avatar processing"""
        
        if processing_id not in self.processing_status:
            return {
                "status": "not_found",
                "progress": 0,
                "message": "Processing ID not found"
            }
        
        return self.processing_status[processing_id]
    
    async def extract_measurements(self, image_file: UploadFile) -> Dict[str, Any]:
        """Extract measurements from uploaded image"""
        
        try:
            image_data = await self._read_and_validate_image(image_file)
            measurements = await self._extract_basic_measurements(image_data)
            
            return measurements
            
        except Exception as e:
            logger.error(f"Error extracting measurements: {str(e)}")
            raise AvatarProcessingException(f"Measurement extraction failed: {str(e)}")
    
    async def validate_measurements(self, measurements: Dict[str, float]) -> Dict[str, Any]:
        """Validate body measurements for consistency"""
        
        warnings = []
        suggestions = []
        corrected_measurements = measurements.copy()
        
        # Basic validation rules
        if "height_cm" in measurements:
            height = measurements["height_cm"]
            if height < 140 or height > 220:
                warnings.append("Height seems unusual (140-220cm expected)")
                if height < 140:
                    corrected_measurements["height_cm"] = 160.0
                    suggestions.append("Adjusted height to 160cm")
                elif height > 220:
                    corrected_measurements["height_cm"] = 180.0
                    suggestions.append("Adjusted height to 180cm")
        
        if "chest_cm" in measurements and "waist_cm" in measurements:
            chest = measurements["chest_cm"]
            waist = measurements["waist_cm"]
            
            if waist > chest:
                warnings.append("Waist measurement larger than chest - this is unusual")
        
        is_valid = len(warnings) == 0
        
        return {
            "is_valid": is_valid,
            "warnings": warnings,
            "suggestions": suggestions,
            "corrected_measurements": corrected_measurements if suggestions else None
        } 