backend/python-ml-service/requirements.txt
# Working ML Service Requirements for Python 3.13
# This skips TensorFlow and other incompatible packages for now

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Basic Image Processing (these work with Python 3.13)
opencv-python>=4.8.0
numpy>=2.0.0
pillow>=10.2.0

# HTTP & API
httpx>=0.28.0
aiofiles>=24.0.0
python-multipart>=0.0.6

# Background Tasks
celery[redis]>=5.5.0
redis>=5.2.0

# Validation
pydantic>=2.11.0

# Development
pytest>=8.4.0
pytest-asyncio>=1.0.0

# Note: TensorFlow and some 3D libraries don't support Python 3.13 yet
# These will be added when they become compatible:
# tensorflow>=2.14.0
# scikit-image>=0.22.0
# scipy>=1.11.0
# trimesh>=4.0.0
# open3d>=0.18.0 